Add a simple string manipulation service
Add a service which can count the number of characters in a string and
capitalize the string sent to it.

The character counting function can be accessed at `/count` by posting
JSON to it. `{"s": "Hello, World!}"`.

The capitalization service can be accessed at `/uppercase` in the same
manner as the counting function.